requirements.txt
apscheduler
python-telegram-bot
python-dotenv
pymongo
requests
web3
pandas
aiohttp
qrcode
base58
solana
solders
anchorpy